Parsing of schedule strings mostly complete.
This version has tests for all of the most recent round of schedule
strings. It's probably going to still have some bugs, but I think if we
wrap it properly in a function that complains when you get the format
wrong and suggests a specific easy to enter format then people will
learn quickly.