Adding support for warning when a game is _close_ to the end time.
In order to support this and have the variable names make more sense I
did some refactoring.

One thing we can't do yet is allow people to post their scheduled games
before the start of the round. Say, for instance, in the case of
pairings being released early.

Not sure how important that is - but it's worth noting.