Initial version of the forwardingMessage command.
This required a bit more refactoring than I originally anticipated. One
of the things I had to do is refactor the slack module to allow it to
be used more than once, which resulted in my having to touch a lot of
code.  Also, I've had to change it so that we are passing the instance
of our bot object into the .on / .hears callback - so I've had to make
that instance work like the botkit.bot.