Fixing references to startPrivateConversation.
Now that the first argument to command callbacks is an instance
of src/slack.js:Bot, not the Botkit.bot, the startPrivateConversation
returns a promise. So this updates ever instance of it to use the
promise contract.

Notably, this simplifies a few spots that were translating the old
callback version into a promise.